+++ apps/connect/CHANGELOG.md
@@ -1,3 +1,35 @@
+## 5.0.0-staging.18 (2025-09-29)
+
+### 🚀 Features
+
+- add document state copy button with toast notification ([#1880](https://github.com/powerhouse-inc/powerhouse/pull/1880))
+
+### ❤️ Thank You
+
+- Guillermo Puente Sandoval @gpuente
+
+## 5.0.0-staging.17 (2025-09-25)
+
+### 🩹 Fixes
+
+- **connect:** added switchboard push listener ([39e8660f5](https://github.com/powerhouse-inc/powerhouse/commit/39e8660f5))
+
+### ❤️ Thank You
+
+- Frank
+
+## 5.0.0-staging.16 (2025-09-25)
+
+### 🚀 Features
+
+- **connect:** open remote drive ([e0a1c43a6](https://github.com/powerhouse-inc/powerhouse/commit/e0a1c43a6))
+- remove hardcoded Vetra drive ID from document editors ([#1878](https://github.com/powerhouse-inc/powerhouse/pull/1878))
+
+### ❤️ Thank You
+
+- Frank
+- Guillermo Puente Sandoval @gpuente
+
 ## 5.0.0-staging.15 (2025-09-24)
 
 This was a version bump only for @powerhousedao/connect to align it with other projects, there were no code changes.