+++ clis/ph-cli/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 5.0.0-staging.18 (2025-09-29)
+
+This was a version bump only for @powerhousedao/ph-cli to align it with other projects, there were no code changes.
+
+## 5.0.0-staging.17 (2025-09-25)
+
+This was a version bump only for @powerhousedao/ph-cli to align it with other projects, there were no code changes.
+
+## 5.0.0-staging.16 (2025-09-25)
+
+### 🚀 Features
+
+- remove hardcoded Vetra drive ID from document editors ([#1878](https://github.com/powerhouse-inc/powerhouse/pull/1878))
+
+### ❤️ Thank You
+
+- Guillermo Puente Sandoval @gpuente
+
 ## 5.0.0-staging.15 (2025-09-24)
 
 ### 🚀 Features