+++ packages/codegen/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 5.0.0-staging.18 (2025-09-29)
+
+This was a version bump only for @powerhousedao/codegen to align it with other projects, there were no code changes.
+
+## 5.0.0-staging.17 (2025-09-25)
+
+This was a version bump only for @powerhousedao/codegen to align it with other projects, there were no code changes.
+
+## 5.0.0-staging.16 (2025-09-25)
+
+### 🚀 Features
+
+- handle replace and duplicate documents in drag and drop ([#1876](https://github.com/powerhouse-inc/powerhouse/pull/1876))
+
+### ❤️ Thank You
+
+- Guillermo Puente Sandoval @gpuente
+
 ## 5.0.0-staging.15 (2025-09-24)
 
 This was a version bump only for @powerhousedao/codegen to align it with other projects, there were no code changes.